Adds a CI test that performs the following:
If the CI test is the result of a pull request (PR) it then pulls the current list of nodes from master, determines which nodes have been added and/or changed in the PR, then does a quick test via a call to /info to determine if the node is reachable. If it is reachable, the test passes; otherwise, the test fails.

Note: If a listed node has not been changed or added the test is skipped to prevent test failures for PRs that are not responsible for those nodes.