merge from upstream (#1)
* Update turtlecoin-nodes.json

Ready

* Update turtlecoin-nodes.json

added in canadian turtle public node

* Remove nodes running outdated daemons

Closes #83 and closes #84

* Remove node runnig outdated daemon

* added alphanumeric test on inital character and updated README

* remove nodes that are unreachable

* fixed missing comma

* fixed test call

* Update turtlecoin-nodes.json

* alphabetized

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Add node to list please :)

* alphabetical

* Add alphabetical check

* Correcting alphabetical order per the newly added test

* @zpalmtree: Could maybe print the sorted list to aid in letting the user fix it?

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* remove errant whitespace

* New Italian node

coinpoolit.webhop.me

* Adding HA Turtle Node to nodes list

* Update turtlecoin-nodes.json

changed Canadian Turtle Public Node host from dkc-hvac.com to mine2.live

* add new node (#105)

Update turtlecoin-nodes.json

* change IP to domain name in turtlecoin-nodes.json (#107)

Update turtlecoin-nodes.json

* added blast's node

* changed ip of blast's node

* added new node

* hi please update node details 

thank you :)

* Update turtlecoin-nodes.json (#112)

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* fixes trtl.pro turtle node url

* add btron

* Update turtlecoin-nodes.json

* Add Sopinka's node

* Adding TurtlePay Blockchain Cache "Nodes" (Supported by zedwallet-beta & wallet-api)

* Update turtlecoin-nodes.json (#118)

Enable SSL on TurtleNode.net

* Update turtlecoin-nodes.json (#119)

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* please update, i have changed the ip name thankyou

* Update turtlecoin-nodes.json

* Add CryptoHub Turtle Blockchain Cache (#124)

* Update turtlecoin-nodes.json

added additional node

* fees, fees, fees.......

* please change thank you..

* update

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Node is running 0.12.1 which is multiple required releases behind

Operator may re-list the node after upgrading it.

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

updating with it in alpha order.

* New request for node listing

I have change the order by alphabet. Thanks

* changing domain

domain expired 🐢

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Add Network 7 powered premium node

* Update turtlecoin-nodes.json

* Network7 node subdomain updated

Preparation for HA node(s)

* Add Network7 slow node

This is HA node, but with limited resources and therefore capacity. This comes with a much lower fee, of course.

* Removed whitespace

* forwarded my node

"Dawns turtlenode"

* removed dawnsturtlenode, added Easy Turtle Node

Pardon my latest request spam but I was initially clueless about how to get a node up properly. So I joined discord, the staff there helped me out a lot and I thank them so much for it. So, for this change request, I removed my old non-static node for a properly set up one.
The should be (HOPEFULLY) my last request, as this time I did things properly: dyndns url, port 11898 is forwarded, you can check with "curl easyturtlenode.ns0.it:11898/info".

Tag me in discord @TheCowboy for anything.

* Removed nodes

Removed my nodes, got better things to use the servers on

* New request to add listing

I have added my public node for access

* Remove unresponsive nodes

* Re-added turtlepool node

It appears whoeever checked this node and removed it was getting load-balanced to one I forgot to update - fixed that, and all the rest are fine!

* Placed it in the correct place alphabetically

* Adds TRTL@MEDIATAO.EU

I got you @CryPiTonia

* Adds TurtleNode.org (EU)

* Update turtlecoin-nodes.json

Add Turtlenode - RUSSIA

* Fix alphabatical order

* Removing TurtleNode.io Nodes

Public access to these nodes (and the HA load balances) is shutting down in favor of operating the Blockchain Cache API via the TurtlePay cache.

* Changing to SSL port

Changing subdomain and port to 443

* Just another node

* Update turtlecoin-nodes.json

* Update

Last time, I promise.

* Quotation mark added

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

Added node.

* Update turtlecoin-nodes.json

removed turtlenode.co as a node

* Removed all nodes that have not responded in the last 4+ days and/or appear abandoned as they are still running 0.14.x

* "Turtlenode - RUSSIA" just woke up from its slumber

* re-adding hashvault as it has awaken

* re-adding greywolf's node #167

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Add AAAAAAAAAAAAAAAAAAAA Node (#169)

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

Unable to sustain

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* remove proton node

can't keep the damn thing up so ¯\_(ツ)_/¯

* shells node (#175)

* Update turtlecoin-nodes.json

* Update turtlecoin-nodes.json

* New node

* Update turtlecoin-nodes.json

Wasn't sure if I supposed to add to the top or bottom...hope the node helps.

* Added & updated Cuvéebits TRTL nodes

* Remove port from url

* Update turtlecoin-nodes.json

Change of IP to Shelly

* add .gitignore

* Adds a CI test that performs the following:

If the CI test is the result of a pull request (PR) it then pulls the current list of nodes from master, determines which nodes have been added and/or changed in the PR, then does a quick test via a call to /info to determine if the node is reachable. If it is reachable, the test passes; otherwise, the test fails.

Note: If a listed node has not been changed or added the test is skipped to prevent test failures for PRs that are not responsible for those nodes.

* Travis actually sets TRAVIS_PULL_REQUEST to "false"

* Update turtlecoin-nodes.json

IP update slight name change

* Update turtlecoin-nodes.json